Adding cls token, cls aggregator and rnn aggregator

+++ src/ndsl/module/preprocessor.py
@@ -1,3 +1,4 @@
+import torch
 import torch.nn as nn
 
 class BasePreprocessor(nn.Module):
@@ -15,5 +16,5 @@ class IdentityPreprocessor(BasePreprocessor):
 class CLSPreprocessor(BasePreprocessor):
     def forward(self, src):
         #src with shape [batch_size, seq_len]
-        return torch.cat((torch.zeros(src.shape[0],1), src), dim=1)
+        return torch.cat((torch.zeros(src.shape[0],1).long(), src), dim=1)
         #src with shape [batch_size, seq_len+1]
